refactor(schedule): types separate for item
